+++ lib/wardrobe/plugins/equality.rb
@@ -3,13 +3,13 @@ module Wardrobe
     module Equality
       extend Wardrobe::Plugin
 
-      option :include_in_equality, Boolean, default: true
+      option :exclude_from_equality, Boolean, default: false
 
       module InstanceMethods
         def ==(other)
           return false unless other.class == self.class
           _attribute_store.all? do |_name, atr|
-            return true unless atr.options[:include_in_equality]
+            return true if atr.options[:exclude_from_equality]
             send(atr.name) == other.send(atr.name)
           end
         end